[oracle] Add initial implementation of oracle plugin Related: RHEL-59870, TYCHE-732
